Divided homepage into several files
hompage were having +700 lines so I decided to create additional files to make the project a bit more organized.
